requirements.txt
fastapi==0.116.1
httpx==0.28.1
openai==1.108.1
python-dotenv==1.1.1
uvicorn[standard]==0.35.0
